Use the same branch completion for B (create) as b (switch)
I often create really long branchnames like
"git-am-ignore-whitespace". When I get comments on the series I want
to create a new branch like "git-am-ignore-whitespace-v2", i.e. just
add the "-v2".

The create branch command makes this really tedious, since it doesn't
offer completion. Change that to use the same completion as "b" uses
while switching branches.

The user can then either type in a branch name as before, or use the
tab completion to partially reuse an old name. Of course this makes it
easier to accidentally reuse an existing name, in which case Git will
fail.

Maybe we could do better in that area, e.g. by simply not accepting
the RET key when the branch exists, but I don't know if the API we're
using can support that.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>